#660 - store - split sync module
