+++ core/store/modules/sync/index.js
@@ -1,0 +1,8 @@
+import actions from './actions'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+  actions,
+  mutations
+}
